Rework the functionality with DOM manipulation
Whole functionality reworked with DOM manipulation
and handler functions. Logic for getting a human
choice is reworked since we now get it through a
button input. Added a restart functionality which
enables us to play the game however many times we
want to. Also, added a very simplistic UI interfa-
ce which helps us track the scores and see the
winner.